refactor: rename widget files and update import paths in create product feature
