+++ apps/seller/lib/presentation/screen/create_product/widgets/featured_toggle_section.dart
