+++ apps/seller/lib/presentation/screen/create_product/widgets/product_details_section.dart
