+++ apps/seller/lib/presentation/screen/create_product/widgets/product_info_section.dart
