+++ apps/seller/lib/presentation/screen/create_product/widgets/product_items_section.dart
